ipc: change svc_id / pid type to size_t
This change is in line with the process / service identifiers being
represented as unsigned integers (usize) in the kernel. It also means
that less checks are required in the userspace applications.

+++ examples/rot13_client/main.c
@@ -4,7 +4,7 @@
 
 #include <ipc.h>
 
-int rot13_svc_num = 0;
+size_t rot13_svc_num = 0;
 
 char buf[64] __attribute__((aligned(64)));
